# api/hqbase-mail-api-v2.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HQBase Mail API",
    "version": "2.0.0",
    "description": "Stable v2 mail API for HQBase clients, automations, and agents. Additive fields may be introduced within v2; clients must ignore unknown response fields. Administrative APIs are not part of this contract."
  },
  "servers": [
    {
      "url": "{origin}",
      "variables": {
        "origin": {
          "default": "https://mail.example.com",
          "description": "HQBase origin"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/v2/messages": {
      "get": {
        "summary": "List or search messages",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "One page of messages, newest activity first.",
            "headers": {
              "Link": {
                "required": false,
                "description": "RFC 8288 link to the next page, as `<url>; rel=\"next\"`. The header is absent on the last page.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageSummary"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. `INVALID_LIMIT` when limit is not an integer from 1 to 100. `INVALID_CURSOR` when the cursor is malformed or is not a message cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "listMessages",
        "parameters": [
          {
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["inbox", "sent", "drafts", "archived", "trash", "catchall"]
            },
            "description": "Message folder to list. Drafts are separate resources, not message rows. Use `/api/v2/drafts` to list drafts."
          },
          {
            "in": "query",
            "name": "mailboxId",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Limit results to one accessible mailbox."
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": "Search one literal substring in the subject, sender, To recipients, snippet, and plain-text body. The characters %, _, and \\ are ordinary text."
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            },
            "description": "Maximum number of messages in the page."
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "description": "Opaque cursor from the previous page. Read it from the Link header and send it unchanged."
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:read"]
      }
    },
    "/api/v2/messages/{id}": {
      "get": {
        "summary": "Get a message",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "getMessage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:read"]
      }
    },
    "/api/v2/messages/{id}/thread": {
      "get": {
        "summary": "Get a message thread",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Accessible thread messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDetail"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "getMessageThread",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:read"]
      }
    },
    "/api/v2/messages/{id}/html": {
      "get": {
        "summary": "Get sanitized message HTML",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sanitized HTML",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageHtml"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "getMessageHtml",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "in": "query",
            "name": "loadRemoteImages",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [1]
            },
            "description": "Load remote images for this response only."
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:read"]
      }
    },
    "/api/v2/messages/{id}/inline/{attachmentId}": {
      "get": {
        "summary": "Render a safe inline image",
        "security": [
          {
            "oauth2": ["mail:read"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inline image",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "contentEncoding": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported inline media",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "getInlineAttachment",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "in": "path",
            "name": "attachmentId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:read"]
      }
    },
...
    "/api/v2/messages/{id}/remote-media/trust": {
      "post": {
        "summary": "Trust remote media from this sender",
        "security": [
          {
            "oauth2": ["mail:write"]
          },
          {
            "cookieSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Trust preference",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteMediaTrust"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "trustRemoteMedia",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          }
        ]
      }
    },
    "/api/v2/messages/{id}/{action}": {
      "post": {
        "summary": "Update message state",
        "security": [
          {
            "oauth2": ["mail:write"]
          },
          {
            "cookieSession": []
          },
          {
            "agentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageSummary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient OAuth scope or mailbox access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": ["Messages"],
        "operationId": "updateMessage",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "read",
                "unread",
                "star",
                "unstar",
                "archive",
                "unarchive",
                "trash",
                "restore"
              ]
            }
          }
        ],
        "x-hqbase-agent-capabilities": ["mail:write"]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "Audience-bound OAuth token. Request resource={origin}/api/v2.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/api/auth/oauth2/authorize",
            "tokenUrl": "/api/auth/oauth2/token",
            "refreshUrl": "/api/auth/oauth2/token",
            "scopes": {
              "mail:read": "Read accessible mail",
              "mail:write": "Organize accessible mail",
              "mail:send": "Manage drafts and send mail",
              "offline_access": "Request a refresh token"
            }
          }
        }
      },
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Same-origin HQBase web session. Secure deployments may prefix the cookie name."
      },
      "agentBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "hqb_agent_<secret>",
        "description": "Mailbox agent credential created in Settings > Agents. It works only for the assigned Mail API capabilities and mailbox grant. Each operation lists required capabilities in x-hqbase-agent-capabilities."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "MessageSummary": {
        "type": "object",
        "required": [
          "id",
          "threadId",
          "mailboxId",
          "direction",
          "folder",
          "fromAddress",
          "to",
          "subject",
          "snippet",
          "receivedAt",
          "sentAt",
          "readAt",
          "starredAt",
          "hasAttachments",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "threadId": {
            "type": "string"
          },
          "mailboxId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "direction": {
            "type": "string",
            "enum": ["inbound", "outbound"]
          },
          "folder": {
            "type": "string",
            "enum": ["inbox", "sent", "drafts", "archived", "trash", "catchall"],
            "description": "Stored message folder. Drafts use the separate Draft schema and endpoints; current write paths do not produce message summaries with `folder: drafts`."
          },
          "fromAddress": {
            "type": "string",
            "format": "email"
          },
          "to": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subject": {
            "type": "string"
          },
          "snippet": {
            "type": "string"
          },
          "receivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "sentAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "readAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "starredAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "hasAttachments": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MessageHtml": {
        "type": "object",
        "required": [
          "afterQuotedHtml",
          "hasRemoteImages",
          "html",
          "quotedHtml",
          "remoteMediaTrusted"
        ],
        "properties": {
          "afterQuotedHtml": {
            "description": "Sanitized authored content that follows quoted reply history, or null when absent.",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "hasRemoteImages": {
            "type": "boolean"
          },
          "html": {
            "type": "string"
          },
          "quotedHtml": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "remoteMediaTrusted": {
            "type": "boolean"
          }
        }
      },
      "RemoteMediaTrust": {
        "type": "object",
        "required": ["remoteMediaTrusted"],
        "properties": {
          "remoteMediaTrusted": {
            "type": "boolean",
            "const": true
          }
        }
      },
...
    }
  }
}